requirements.txt
Flask==2.3.3
Pillow==10.0.1
moviepy==1.0.3
openai==1.3.8
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==2.3.7